Remove vestigial parts of the old daemon
