--- aiida/daemon/settings.py
@@ -1,10 +1,0 @@
-
-
-DAEMON_INTERVALS_SUBMIT = 30
-DAEMON_INTERVALS_RETRIEVE = 30
-DAEMON_INTERVALS_UPDATE = 30
-DAEMON_INTERVALS_WFSTEP = 30
-DAEMON_INTERVALS_TICK_WORKFLOWS = 30
-DAEMON_MAX_PARALLEL_PROCESSES = 20
-# Use the new, thread based daemon
-DAEMON_USE_NEW = True
